physical-event-experience/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
google-genai
flask-talisman
flask-wtf
pytest
python-dotenv
